requirements.txt
adafruit-circuitpython-led-animation >= 2.9.0
adafruit-circuitpython-neopixel >= 6.3.11
adafruit-circuitpython-pixel-framebuf >= 1.1.13
aio-pika >= 9.0.0
python-dotenv >= 1.0.0
aiomisc >= 17.0.0
pillow >= 10.3.0